verbos_backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.7.0
django-ipware==7.0.1
djangorestframework==3.15.2
fastapi==0.115.8
filelock==3.16.1
h11==0.14.0
idna==3.10
pillow==11.2.1
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-ipware==3.0.0
qrcode==8.1
reportlab==4.4.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.45.3
svgwrite==1.4.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
virtualenv==20.26.6